Project Scope
A database design generally project starts with requirements. You analyze the requirements to identify the business rules, entities, and relationships. In this case, however, we are using snapshots from a functional website (https://www.fiverr.com/) to understand the entities and business rules so that we can reverse engineer the database design.

Using information from the functional website (https://www.fiverr.com/) to understand the entities and business rules to reverse engineer the database design using MYSQL

Due to the nature of the project, i am unable to post the entire solution/project. Including few screenshots for reference.